+++ CHANGELOG.md
@@ -12,7 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Entry types for known storage models
 - `filter` implementation for storage models
 - `sorter` implementation for storage models
+- `group_with_members` struct as a replacement for `group`
+- static variables for group roles
 
 ### Changed
 
 - Bump `ic-cdk` crate to version `0.15`
+-